@@ -1,6 +1,6 @@
 {
     "name": "react-typescript-webpack-starter",
-    "version": "2.1.0",
+    "version": "3.0.0",
     "description": "A starter project for using React, TypeScript, SCSS and Webpack",
     "main": "index.js",
     "scripts": {
@@ -16,29 +16,29 @@
     "author": "Grant Bartlett",
     "license": "MIT",
     "dependencies": {
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2"
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0"
     },
     "devDependencies": {
-        "@types/react": "^17.0.5",
-        "@types/react-dom": "^17.0.3",
-        "clean-webpack-plugin": "^3.0.0",
-        "copy-webpack-plugin": "^8.1.1",
-        "css-loader": "^5.2.4",
+        "@types/react": "^18.0.15",
+        "@types/react-dom": "^18.0.6",
+        "clean-webpack-plugin": "^4.0.0",
+        "copy-webpack-plugin": "^11.0.0",
+        "css-loader": "^6.7.1",
         "file-loader": "^6.2.0",
-        "html-webpack-plugin": "^5.3.1",
-        "mini-css-extract-plugin": "^1.6.0",
-        "postcss": "^8.2.14",
-        "postcss-loader": "^5.2.0",
-        "postcss-preset-env": "^7.2.3",
-        "sass": "^1.32.12",
-        "sass-loader": "^11.0.1",
-        "style-loader": "^2.0.0",
-        "terser-webpack-plugin": "^5.1.1",
-        "ts-loader": "^9.1.2",
-        "typescript": "^4.2.4",
-        "webpack": "^5.36.2",
-        "webpack-cli": "^4.7.0",
-        "webpack-dev-server": "^4.7.3"
+        "html-webpack-plugin": "^5.5.0",
+        "mini-css-extract-plugin": "^2.6.1",
+        "postcss": "^8.4.14",
+        "postcss-loader": "^7.0.1",
+        "postcss-preset-env": "^7.7.2",
+        "sass": "^1.53.0",
+        "sass-loader": "^13.0.2",
+        "style-loader": "^3.3.1",
+        "terser-webpack-plugin": "^5.3.3",
+        "ts-loader": "^9.3.1",
+        "typescript": "^4.7.4",
+        "webpack": "^5.73.0",
+        "webpack-cli": "^4.10.0",
+        "webpack-dev-server": "^4.9.3"
     }
 }
